Mv rawio test folder. (#707)
* Mv rawio test folder.

* Modify setup

* pep8

* oups

* oups
