+++ neo/test/rawiotest/rawio_compliance.py
