+++ neo/test/rawiotest/test_blackrockrawio.py
@@ -9,7 +9,7 @@ from __future__ import unicode_literals, print_function, division, absolute_impo
 import unittest
 
 from neo.rawio.blackrockrawio import BlackrockRawIO
-from neo.rawio.tests.common_rawio_test import BaseTestRawIO
+from neo.test.rawiotest.common_rawio_test import BaseTestRawIO
 
 import numpy as np
 from numpy.testing import assert_equal
