+++ neo/test/rawiotest/test_neuroscoperawio.py
@@ -6,7 +6,7 @@ from __future__ import unicode_literals, print_function, division, absolute_impo
 import unittest
 
 from neo.rawio.neuroscoperawio import NeuroScopeRawIO
-from neo.rawio.tests.common_rawio_test import BaseTestRawIO
+from neo.test.rawiotest.common_rawio_test import BaseTestRawIO
 
 
 class TestNeuroScopeRawIO(BaseTestRawIO, unittest.TestCase, ):
