+++ neo/test/rawiotest/test_winedrrawio.py
@@ -5,17 +5,16 @@ from __future__ import unicode_literals, print_function, division, absolute_impo
 
 import unittest
 
-from neo.rawio.plexonrawio import PlexonRawIO
+from neo.rawio.winedrrawio import WinEdrRawIO
+from neo.test.rawiotest.common_rawio_test import BaseTestRawIO
 
-from neo.rawio.tests.common_rawio_test import BaseTestRawIO
 
-
-class TestPlexonRawIO(BaseTestRawIO, unittest.TestCase, ):
-    rawioclass = PlexonRawIO
+class TestWinEdrRawIO(BaseTestRawIO, unittest.TestCase, ):
+    rawioclass = WinEdrRawIO
     files_to_download = [
-        'File_plexon_1.plx',
-        'File_plexon_2.plx',
-        'File_plexon_3.plx',
+        'File_WinEDR_1.EDR',
+        'File_WinEDR_2.EDR',
+        'File_WinEDR_3.EDR',
     ]
     entities_to_test = files_to_download
